fix: Add the missing length for the Multihash checksum
A multihash pattern is as follows:

```
{code}{len}{hash}
```

where `code` and `len` are variable unsigned integers (i.e. each one
takes one or more bytes).

The current case, Blake3, requires one byte each so the algorithm is a
simple slicing operation. if we introduce multiple algorithms we
will need a full variable integer implementation but given what we know
we want right now, this is unlikely to happen.

Signed-off-by: Arnau Siches <asiches@gmail.com>